backend/requirements.txt
Django~=3.2.9
phonenumbers~=8.12.39
djangorestframework~=3.12.4
requests~=2.25.1